Fixes bug for two-qubit gate application on tensors on the far left of the MPS.
The bug was trying to use the naming convention to determine which are the left and right connected edges.
If the newly formed MPS tensor was the farthest left in the MPS, it didn't have a left neighbor with the name "q{integer}", so the determination of what was left/right would fail.

Fixed by adding a check and added unit tests.